test(mssql): run bulk copy tests without a server and cover parquet in the load suite
